producer/requirements.txt
kafka-python==2.0.2
websockets==10.2